refactor: 모든 shared/ui import를 barrel export 경로로 통일
기존 직접 경로(@/shared/ui/Footer 등)를 barrel export(@/shared/ui)로
변경하여 import 경로를 일관되게 정리.

- ClientLayout, layout, page, ClientPagination, ImageWithModal 업데이트
- barrel export에 Header 플러그인 팩토리 함수 추가
