+++ src/shared/ui/index.ts
@@ -10,3 +10,10 @@ export { Footer } from "./Footer";
 export { Pagination } from "./Pagination";
 export { default as BottomNavigation } from "./BottomNavigation";
 export type { HeaderPlugin, HeaderSectionProps } from "./Header/types";
+export {
+  createLogoPlugin,
+  createNavigationPlugin,
+  createSearchPlugin,
+  createMobileSearchPlugin,
+  createThemeTogglePlugin,
+} from "./Header";
